Remove slow assertion from module-name util
Was playing around and profiling the new npm5 release and it's still a drop sluggish in cases where it could be much faster (like when it already has package-lock.json). I can provide benchmarks if necessary to get this landed, just point me in the right direction. In my tests, it speeds up install by approx 25%!

The assertion appears to just be an abundance of caution, as I checked all call sites and it seems like an object is guaranteed in each of those cases. Let me know your thoughts, thanks for considering the patch and thanks for everything you all do!

P.S. Excuse the `patch-1` branch. I tested this on my computer but am recreating the pr on my phone!